requirements.txt
asgiref==3.4.1
dj-static==0.0.6
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
